tests/wpt/web-platform-tests/2dcontext/the-canvas-state/2d.state.saverestore.shadowOffsetX.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.state.saverestore.shadowOffsetX</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.shadowOffsetX</h1>
<p class="desc">save()/restore() works for shadowOffsetX</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() works for shadowOffsetX");
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.shadowOffsetX;
ctx.save();
ctx.shadowOffsetX = 5;
ctx.restore();
_assertSame(ctx.shadowOffsetX, old, "ctx.shadowOffsetX", "old");

// Also test that save() doesn't modify the values
ctx.shadowOffsetX = 5;
old = ctx.shadowOffsetX;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against 5
ctx.save();
_assertSame(ctx.shadowOffsetX, old, "ctx.shadowOffsetX", "old");
ctx.restore();


});
</script>
